+++ common/src/main/java/com/tendai/common/source/model/Playlist.kt
@@ -4,5 +4,5 @@ package com.tendai.common.source.model
 data class Playlist(
     var playlistId: Long = -1,
     var playlistName: String = "",
-    var numberOfTracks: Int = -1
+    var numberOfTracks: Long = -1
 )
